Move run logic in core Pipeline
